# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f46db3167541c707d179879
  title: appvaa
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/cities/{id}":
    delete:
      x-skaffolder-id: 5f46eb1367541c707d1799c2
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e56767541c707d17994e
      x-skaffolder-resource: Cities
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Cities
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eb1367541c707d1799bc
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e56767541c707d17994e
      x-skaffolder-resource: Cities
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Cities
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cities"
    post:
      x-skaffolder-id: 5f46eb1367541c707d1799bd
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e56767541c707d17994e
      x-skaffolder-resource: Cities
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Cities
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cities"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cities"
...
  "/classes/{id}":
    delete:
      x-skaffolder-id: 5f46eb1867541c707d1799d2
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e59767541c707d179956
      x-skaffolder-resource: Classes
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Classes
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eb1767541c707d1799c7
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e59767541c707d179956
      x-skaffolder-resource: Classes
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Classes
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classes"
    post:
      x-skaffolder-id: 5f46eb1767541c707d1799c8
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e59767541c707d179956
      x-skaffolder-resource: Classes
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Classes
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Classes"
  "/classes/findBy_club/{key}":
    get:
      x-skaffolder-id: 5f46ead567541c707d179990
      x-skaffolder-name: findBy_club
      x-skaffolder-id-resource: 5f46e59767541c707d179956
      x-skaffolder-resource: Classes
      x-skaffolder-crudAction: findBy_club
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_club
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_club/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_club
      tags:
        - Classes
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Classes"
  "/classes/findBy_student/{key}":
    get:
      x-skaffolder-id: 5f46ead567541c707d179992
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 5f46e59767541c707d179956
      x-skaffolder-resource: Classes
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - Classes
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Classes"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Cities:
      x-skaffolder-id: 5f46e56767541c707d17994e
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e56767541c707d17994c
      x-skaffolder-url: /cities
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5f46e56767541c707d17994d
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    Classes:
      x-skaffolder-id: 5f46e59767541c707d179956
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e59767541c707d179951
      x-skaffolder-url: /classes
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        days:
          type: number
          x-skaffolder-id-attr: 5f46e59767541c707d179952
          x-skaffolder-type: Number
        vacancies:
          type: number
          x-skaffolder-id-attr: 5f46e59767541c707d179953
          x-skaffolder-type: Number
      x-skaffolder-relations:
        _club:
          x-skaffolder-id: 5f46e59767541c707d179954
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e59767541c707d179951
          x-skaffolder-ent2: 5f46e51a67541c707d17992b
        _student:
          x-skaffolder-id: 5f46e59767541c707d179955
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e59767541c707d179951
          x-skaffolder-ent2: 5f46e51a67541c707d17992b
      required:
        - _id
...